requirements.txt
astroid==2.11.6
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
dill==0.3.5.1
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.3
httplib2==0.20.4
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
oauthlib==3.2.0
platformdirs==2.5.2
protobuf==4.21.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.14.4
pyparsing==3.0.9
python-dotenv==0.20.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
tomli==2.0.1
tomlkit==0.11.0
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.14.1
yapf==0.32.0
